[GEOT-7764] Migrate GeoTools to Java 17
* Upgrade build for Java 17 as base Java version
  * Configure maven-enforcer-plugin to require Java 17
  * Configure maven-enforcer-plugin to require maven 3.8+ as done in GeoServer.
  * Remove cobertura-maven-plugin (modules/extension/xsd/**), as it is not compatible with Java 11+
  * Update maven-javadoc-plugin configuration to use legacyMode=true and not fail due to module-info.java not present.
* Upgrade github actions workflows to use Java 17, remove the linux_jdk11.yml workflow.
* Centralize plugin versions in the root pom's pluginManagement section
* Update docs for Java 17
* Add .mvn/jvm.config with necessary --add-exports and --add-opens for internal jdk.compiler packages.
  This ensures that build-time tools and compiler plugins (e.g., Error Prone, and others interacting with javac internals) can
  function correctly under Java 17's module system.
* Plugin version upgrades:
  * jacoco 0.8.7 -> 0.8.13
  * maven-checkstyle-plugin 3.3.0 -> 3.6.0
  * maven-compiler-plugin 3.11.0 -> 3.14.0
  * maven-javadoc-plugin 3.6.3 -> 3.11.2
  * maven-surefire-plugin 2.22.2 -> 3.5.3
  * maven-failsafe-plugin 2.22.2 -> 3.5.3
* Upgrade errorprone to work with Java 17:
  * add suggested compiler args `-Xep:DeadException:WARN -Xep:GuardedBy:OFF`
  * Add additional errorprone exclusions and fix trivial errors
  * Use explict charset reading streams (IntelliJ)
  * Inner class can be static, by IntelliJ
  * Removing unnecessary parenthesis (intellij)

* Fix usage of varargs in GridCoverageReader/Writer

  GridCoverageReader.read() and GridCoverageWriter.write() were declared
  to receive `GeneralParameterValue... parameters` varargs but most
  classes in their hiearchy were still using `GeneralParameterValue[]`,
  which caused errors when compiling with `-Werror` and Java 17.

  Update to replace calls with null argument like `reader.read(null)` by
  `reader.read()` and avoid the "inon-varargs call of varargs method with
  inexact argument type for last parameter" errors.

  Update ImageMosaicReader, WorldImageWriter, and GrassCoverageWriter, to
  be aware that a client call without values in the varargs parameter
  means an empty array. All other implementations didn't need it because
  the idiom `if(parameters != null){for (GeneralParameterValue param :
  parameters)...` doesn't need updating.

* Fix new -Werror compilation errors
  * Remove `strictfp` keywords: as of release 17, all floating-point expressions are evaluated strictly and 'strictfp' is not required
  * java.security.AccessController in java.security has been deprecated and marked for removal
  * non-varargs call of varargs method with inexact argument type for last parameter
  * deprecated item is not annotated with @Deprecated
  * deprecated item is not annotated with @Deprecated
  * unreachable catch clause

* Update the build/maven/jjtree-javacc plugin to add @SuppressWarnings in generated java files
  Upgrade net.java.dev.javacc:javacc to 7.0.13 and update JJTreeJavaCC.java to add @SuppressWarnings("unchecked")
  to  generated parsers so they work with javac -Werror

* Relax tolerance for GridCoverageRendererTest.testRenderRadarOnAzeq() from 20 to 50 pixels

  Update GridCoverageRendererTest.testRenderRadarOnAzeq() with a 50 pixel
  tolerance when comparing the expected and the rendered images, due to
  differences between rendering on Linux and MacOs with Java 17.

+++ modules/unsupported/ogr/ogr-jni/src/main/java/org/geotools/data/ogr/jni/JniOGR.java
@@ -368,7 +368,7 @@ public class JniOGR implements OGR {
 
     @Override
     public boolean LayerDeleteFeature(Object layer, long fid) {
-        return ((Layer) layer).DeleteFeature((int) fid) == 0;
+        return ((Layer) layer).DeleteFeature(fid) == 0;
     }
 
     @Override
